[Liftoff] Use better value for "undefined Register"
There is no "no_reg" LiftoffRegister, so we sometimes have to
initialize it to an arbitrary value, because it will be overwritten
later anyway. One solution would be to wrap it in base::Optional, but
that does not really reflect the right semantics, since it's not
optional, but just set at a later point.
So far, we used the register with code 0, but on some platforms (e.g.
mips), register 0 is not a Liftoff cache register. So just use the
first gp cache register instead.

R=ahaas@chromium.org

Bug: chromium:883280
Change-Id: I157384ad5a9ca6a20aa198f3217a7d754dc53e30