[objects] Change JSArrayBuffer::byte_length to uintptr_t.
Previously the [[ArrayBufferByteLength]] internal field was represented
as a boxed number (i.e. either Smi or HeapNumber) in safe integer range.
This is the first step to change the representation of all the array
buffer and array buffer view length/offset fields to unboxed integers,
to eventually support the full range of 4GiB (and potentially even more)
for typed arrays and array buffers. This will allow WebAssembly memories
with 4GiB to be usable.

R=mstarzinger@chromium.org, titzer@chromium.org, yangguo@chromium.org

Bug: v8:7881, v8:8015, v8:8171
Change-Id: Ic6c6c8fe087afee898254cd903e82a55bfc173a9