Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > Hi Deepti,
> > > 
> > > Do you actually also have positive/negative tests, where you test that all possible aligned memory accesses do actually work?
> > > 
> > > Cheers, Andreas
> > 
> > I'm not sure what you mean, could you elaborate? Are you looking for tests of all possible combinations of alignment? I've added tests for individual types, and operations that share code in the compiler i.e. if multiple operations go through the same macro/method I've added a representative test.
> 
> Sorry, I should have been more clear. I mean that in the test you added you check that accesses with invalid alignment cause a trap. However, afaict you do not test if accesses with valid alignment do not cause a trap. I wonder if you put tests like that somewhere else, or if we should add such tests.

All the tests that currently pass test for variable offsets till the bounds, so these are positive tests. For example, if you look at the first GetAtomicBinOpFunction, it takes offset as an argument, so most of the tests except the last ones are positive tests that would fail the runtime trap. If an incorrect alignment was added, this would be a validation failure, that's common to all memory access operations so I don't think it needs special tests here. Let me know if you think there's anything else missing from this set.

Patch-set: 4