Generalize ExtractFixedArray's branch for Smis or Objects.
This CL extracts the branch into a new function. Previously, the branch
is only used to copy a FixedArray to a new FixedArray. The new function
generalizes this to allow copying from a FixedDoubleArray to a
FixedArray also. This function will be useful in a follow-up CL to copy
a FixedDoubleArray with holes into a FixedArray where holes are replaced
by undefined.

Bug: chromium:881273

Change-Id: I8a0e5f933fc152a12d67810f4cbcfdce094d44af