Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > Patch Set 9:
> > 
> > > Patch Set 9:
> > > 
> > > Can you explain the motivation for making the deleted special members public a bit more. I think it is good for consistency to have the {DISALLOW_COPY_AND_ASSIGN} macro be the last thing in any class/struct declaration. This is also being checked by a presubmit check which is failing on this CL.
> > 
> > +1
> 
> Your right, for consistency moving everything down is better. I'll fix that.
> 
> The reasoning behind making them public is twofold: First, by deleting these members we change the behavior(copy semantics) of the class, which influences usage of the class and therefore should be "communicated" to the user via the public interface of a class.
> Second, for some situations/compilers this can produce better error messages. You get `Calling explicitly deleted member function foo` instead of linker errors or foo was not defined.
> 
> The origin of making these members private comes from pre c++11 where there was no `= delete` and one had to declare but not define these functions and make them private but due to `delete` this is no longer required.

I see your second point here about the better error message, that could indeed offer slightly faster/better feedback to the user of the code. As for the second point, there not being any explicit copy constructor in the public section of the interface at least warrants a full read-through of the class declaration including the private fields IMHO for the user to understand what implicit copying entails.

Also, as you mentioned offline, even the style guide mentions that the deletion should happen in the public section[1].

My biggest concern (and source of push-back) against this refactoring is consistency. Taking a glance at the rest of Chromium[2] shows that almost all placements of the {DISALLOW_COPY_AND_ASSIGN} macro are at the end of the private section. This would make V8 inconsistent to the rest of Chromium, and make files that have been copied to V8 (e.g. src/base/adapters.h) diverge even further. I am not sure the benefits of this refactoring outweigh this inconsistency. But I a happy to be convinced otherwise in case other reviewers are in favor of this refactoring.

And for the sake of completeness, I would definitely don't like us to have a trailing public section in the end of all non-copyable classes just to keep the macro at the end and make clang-tidy happy. That would be the worst of both worlds IMHO.

[1] https://google.github.io/styleguide/cppguide.html#Copyable_Movable_Types
[2] https://cs.chromium.org/search/?q=disallow_copy_and_assign&sq=package:chromium&type=cs

Patch-set: 9
Reviewer: Gerrit User 1118575 <1118575@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>