[heap profiler] Refactor: Replace indices with HeapEntry*
We used to accumulate entries in a std::vector that had two major drawbacks:
1. It copies all the entries on resize
2. Entry pointers are not stable due to reallocations.

The second property forced us to use indices to entries as an invariant everywhere,
and then translate them to pointers.

The patch replaces backing storage for entries with std::deque that does not
have these two limitations. That makes code cleaner and removes danger of
holding pointer to a heap entry that has been moved.

Change-Id: I0c176f66711d45e2f59d527f3133a1afbf825ec3