Update patch set 9
Patch Set 9:

Indeed we have the convention that DISABLE_... is the last line. And now we want to make it public. IMO it's ugly to add another "public:" block at the end of the class - this also violates the convention that publics are at the beginning of the class.

How about changing the convention and making DISABLE_... the first line of the class, just after the opening of the public block: ? Wdyt all (esp. Ross, Michael S.)?

Patch-set: 9
Reviewer: Gerrit User 1115961 <1115961@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>