Add fast path for spreading strings.
This improves the performance on strings of IterableToListWithSymbolLookup,
which implements the CreateArrayFromIterable bytecode. The fast path is
only taken if the Symbol.iterator of String.prototype, the Symbol.iterator
of the target string, and the String Iterator prototype are untouched.

This makes IterableToListWithSymbolLookup on par with the optimizations
done for string iterator.

Bug: chromium:881273
Change-Id: Ic8d8619da2f2afcc9346203613a844f62653fd7a