Add fast path for spreading strings.
This improves the performance on strings of IterableToListWithSymbolLookup,
which implements the CreateArrayFromIterable bytecode. The fast path is
only taken if the lookup of Symbol.iterator on the input string is the
original iterator function, and the String Iterator prototype are
untouched.

This brings spreading of strings closer to the performance of the string
iterator performance.
(see https://docs.google.com/document/d/13z1fvRVpe_oEroplXEEX0a3WK94fhXorHjcOMsDmR-8/).

Bug: chromium:881273
Change-Id: Ic8d8619da2f2afcc9346203613a844f62653fd7a