[parser] Name outer promise ".promise" in the desugaring.
The Parser inserts a temporary .promise variable to hold the outer
promise for async functions, but doesn't assign an actually visible
name to it. This change puts the name ".promise" on it (similar to
what we do for other special variables like .result), which makes it
easier to debug this (for V8 developers). It's not observable to
regular developers.

R=adamk@chromium.org

Bug: v8:7522, v8:8015
Change-Id: I388352ad931a49102cb5afe350314337a505a150