[ia32,root] Enable PreserveRoot register configuration
The configuration is enabled depending on whether the call descriptor flag
CallDescriptor::kCanUseRoots is set. This allows Code::Kind::STUB code to
not support the root register (useful for multi-return fuzzer, for example).

It also turns on the restricted configuration for the JS_TO_WASM stubs, because
they are basically JSFunctions, and for the WASM_TO_JS stubs, because they will
contain a call to a JSFunction and need to reset the root register there (which
is difficult if all 6 registers are allocated to arguments, context, and function
address for a call via register).

This CL breaks the tests because the WASM_TO_JS stub now can't contain a call to
the arguments adaptor trampoline anymore: That one needs six registers. However,
once the arguments adaptor trampoline is isolate independent, those calls will start
working once more, because we can inline the trampoline then.

Change-Id: I223a3a15490dcc48b80c10629e966d6e17a41d43
Bug: v8:6666