Add iterator protectors for JSMap and JSSet.
The JSMap iterator protector protects the original iterating behavior
of .keys() and .values(). It does not protect the behavior of .entries()
or [Symbol.iterator](). The protector is invalidated when (1) .next is
changed on an object whose lookup of next leads up to %MapIteratorPrototype%
or (2) Symbol.iterator is changed on an iterator object whose lookup of
Symbol.iterator leads up to %IteratorPrototype%.

The JSSet iterator protector protects the original iterating behavior of
.key(), .values(), and [Symbol.iterator](). It does not protect the
behavior of .entries(). The protector is invalidated when (1) .next is
changed on an object whose lookup of next leads up to %SetIteratorPrototype%
or (2) Symbol.iterator is changed on an object whose lookup of Symbol.iterator
leas up to %SetPrototype% OR %IteratorPrototype%.

Bug: v8:7980
Change-Id: I54ad6e4c7f19ccc27d7001f6c4b6c8d6ea4ee871