Add iterator protectors for JSMapIterator/JSSet/JSSetIterator.
The MapIterator protector protects the original iteration behaviors of
JS_MAP_KEY_ITERATOR_TYPE (Map.prototype.keys()) and
JS_MAP_VALUE_ITERATOR_TYPE (Map.prototype.values()). It does not protect
the behavior of JS_MAP_KEY_VALUE_ITERATOR_TYPE, which is the behavior of
Map.prototype.entries() and Map.prototype[Symbol.iterator](). The protector
is invalidated when:
* The 'next' property is set on an object where the property holder is the
  %MapIteratorPrototype% (e.g. because the object is that very prototype).
* The [Symbol.iterator] property is set on an object where the property
  holder is the %IteratorPrototype%. Note that this also invalidates the
  SetIterator protector (see below).

The SetIterator protector protects the original iteration behavior of
JS_SET_VALUE_ITERATOR_TYPE, which is the behavior of Set.prototype.keys(),
Set.prototype.values(), and Set.prototype[Symbol.iterator]() of JSSet.
It does not protect the behavior of JS_SET_KEY_VALUE_ITERATOR_TYPE, which
is the behavior of Set.prototype.entries(). The protector is invalidated
when:
* The 'next' property is set on an object where the property holder is the
  %SetIteratorPrototype% (e.g. because the object is that very prototype).
* The [Symbol.iterator] property is set on an object where the property
  holder is the %SetPrototype% OR %IteratorPrototype%. This means that
  setting Symbol.iterator on a MapIterator object can also invalidate the
  SetIterator protector, and vice versa, setting Symbol.iterator on a
  SetIterator object can also invalidate the MapIterator. This is an over-
  approximation for the sake of simplicity.

Bug: v8:7980
Change-Id: I54ad6e4c7f19ccc27d7001f6c4b6c8d6ea4ee871