[turbofan] Fix representation selection of CheckFloat64Hole.
If the CheckFloat64Hole becomes a noop after RETYPE (because the
feedback type is already Number), we don't pass a restriction type.
But since the previous pass already passed a restriction type
Number, this doesn't work anymore. This previous passing of type
Number is unnecessary anyways, and probably wrong too.

Bug: chromium:895199
Change-Id: I96a949ba35db1e6d35abedddc4507c101d95b716