Update patch set 3
Patch Set 3:

It sounds like this patch takes the implementation from previously being complaint to the specification at master and instead implements the suggestion from @anba on the thread. As I say on that thread, I personally like this suggestion better. I didn't take this issue to the committee because it got some resistance there and it didn't seem like a big deal, but if this is a point of semantics that you feel strongly about, I'll take it to TC39 in November.

Too bad there aren't test262 tests covering this case one way or the other, cc @ms2ger.

Patch-set: 3
CC: Gerrit User 1257330 <1257330@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1129576 <1129576@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>