Update patch set 3
Patch Set 3:

Is it possible to emit this flood of "Attempted to kill [...]" messages only when some debug/verbose flag is given? Now I always have to scroll up in my shell to see the test results after aborting the test run, because the last 100 lines or so are noise.

Patch-set: 3
CC: Gerrit User 1119597 <1119597@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>