Update patch set 3
Patch Set 3:

You don't need to make a change separately. I'm already doing it as a part of ICU 63 update. Well, it's harmless to do it separately, too :-). But, there's another test adjustment that is necessary so that I can do this along with that.   Anyway, either way is fine. 

The reason  this change is necessary ('loose >= normal') is that I now include loose rules in ICU 63 (the rule has been there but I didn't include it in Chrome's ICU up to 62 out of the data size concern; loose is rarely used in CSS and without Ecma 402 Intl.BreakIterator, there's hardly any need for it). In ICU 63, Finnish tailoring has been absorbed into the root locale (I filed that request many years ago against ICU and it's finally resolved in ICU 63 !) making room for  the loose rule for the root locale.

Patch-set: 3
Reviewer: Gerrit User 1151758 <1151758@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>