[esnext] implement private methods
This patch is a initial implementation of the stage 3 proposal
https://tc39.github.io/proposal-private-methods

- Adds a --harmony-private-methods flag
- Parse the private methods/accessors
- Implement the semantics of private methods/accessors

This patch also renames a few variables and flag names so that the
code can be reused by private methods implementation.

- Rename private fields to private names where it makes sense
- Rename class fields to class elements where it makes sense

The design doc is in
https://docs.google.com/document/d/1T-Ql6HOIH2U_8YjWkwK2rTfywwb7b3Qe8d3jkz72KwA/edit?usp=sharing

The private methods are added to the instances in the
instance_elements_initializer function through the
InitializeClassElementsStatement (previously
instance_fields_initializer and VisitInitializeClassFieldsStatement,
renamed to echo the spec).

Bug: v8:8330
Change-Id: I55f0ebdf870f7807315a3ce8e3e7e38713233c4d