Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> Just skimmed this and mostly looks good! Let me know if you'd like me to have a more detail review
> 
> - Only class fields have an initializer, private methods don't. So potentially the renaming can be more contained -- for ex - I don't see why the function name is now class_elements_initializer, if this can only contain class fields.
> - It'd be great to split out the renaming into a separate CL.

Thanks, I am currently working on fixing tests and adding more tests, there are also a few edge cases with proxies to fix as well.

> I don't see why the function name is now class_elements_initializer, if this can only contain class fields.

So currently with a class like this

class A {
  #x = 1;
  get #foo() {
    return 1;
  }
}

this patch generates

[generated bytecode for function: <instance_elements_initializer>]
Parameter count 1
Frame size 24
   12 E> 0x76082ca0352 @    0 : a5                StackCheck
         0x76082ca0353 @    1 : 1b 04             LdaImmutableCurrentContextSlot [4]
         0x76082ca0355 @    3 : 26 fa             Star r1
   17 S> 0x76082ca0357 @    5 : 0c 01             LdaSmi [1]
         0x76082ca0359 @    7 : 26 f9             Star r2
         0x76082ca035b @    9 : 27 02 fb          Mov <this>, r0
         0x76082ca035e @   12 : 61 d6 00 fb 03    CallRuntime [AddPrivateField], r0-r2
         0x76082ca0363 @   17 : 1b 05             LdaImmutableCurrentContextSlot [5]
         0x76082ca0365 @   19 : 26 fa             Star r1
   22 S> 0x76082ca0367 @   21 : 81 00 00 02       CreateClosure [0], [0], #2
         0x76082ca036b @   25 : 26 f9             Star r2
         0x76082ca036d @   27 : 61 d8 00 fb 03    CallRuntime [AddPrivateGetter], r0-r2
         0x76082ca0372 @   32 : 0d                LdaUndefined
   52 S> 0x76082ca0373 @   33 : a9                Return

Which basically echos InitializeInstanceElements in the spec: https://tc39.github.io/proposal-private-methods/#initialize-public-instance-elements

Hence the naming. But I am conflicted on whether to split the field installation and the method installation into different functions, or just put them together in one instance_elements_initializer - as shown above currently it does not do exactly what the spec says, the accessor is added after the field instead of before it. Do you think it's better to split into two functions, or loop through the elements twice in the same intializer as described in the spec?

> - It'd be great to split out the renaming into a separate CL.

Agreed. Was thinking about that but I was afraid simply renaming a bunch of things without a patch showing the follow-on would be hard to review if they all make sense - like the initializer renaming you've pointed out.

Patch-set: 11
Reviewer: Gerrit User 1288325 <1288325@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false